requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
cffi==1.11.5
confusable-homoglyphs==3.0.0
coverage==4.5.1
cryptography==2.2.2
decorator==4.3.0
Django==2.0.4
django-appconf==1.0.2
django-compressor==2.2
django-multiselectfield==0.1.8
django-registration==2.4.1
django-sass-processor==0.7
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.13
flake8==3.5.0
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
libsass==0.14.5
MarkupSafe==1.0
mccabe==0.6.1
model-mommy==1.5.1
paramiko==2.4.1
parso==0.2.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyNaCl==1.2.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
